Made discs become drawn when the board is tapped.
Also changed the board view delegate.
Also changed the board view to store points instead of the disc coordinate class.

+++ CrokinoleTracker/CrokinoleTracker/Round.h
@@ -26,8 +26,8 @@
 @property (nonatomic, retain) NSNumber *playerTwo5s;
 @property (nonatomic, retain) Game *game;
 
-- (void)addDiscPositionAtX:(int)x
-                         y:(int)y;
+- (void)addDiscPositionAtX:(CGFloat)x
+                         y:(CGFloat)y;
 - (int)scoreForPlayer:(Player *)player;
 - (int)playerOneScore;
 - (int)playerTwoScore;
